Moved NetworkX implementatioin to separate file.
